test/tizen/offloadMediaDevicesTC/testcases/test_enumerateDevices.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8" />
  <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0">
  <meta name="timeout" content="long">
  <title>Offload MediaDevices Test Case</title>
  <link rel="stylesheet" type="text/css" href="../css/style.css" />
  <link rel="author" title="Dr Alex Gouaillard" href="mailto:agouaillard@gmail.com"/>
  <link rel="help" href="https://w3c.github.io/mediacapture-main/#enumerating-devices">
  <meta name='assert' content='Check that the enumerateDevices() method is present.'/>
  <script src="../resources/testharness.js"></script>
  <script src="../resources/testharnessreport.js"></script>
  <script src="http://127.0.0.1:9559/offload.js"></script>
  <script src="../js/test.js"></script>
</head>
<body>
<p>offloaded enumerateDevices: test that enumerateDevices is present and works properly</p>
<div>
  <p>Test steps:</p>
  <ol>
    <li>Test passes if all the test entries below show "pass".</li>
  </ol>
</div>
<video id="video" autoplay controls></video>
<script>
  promise_test(async () => {
    await navigator.mediaDevices.getUserMedia({ video : true });
    const deviceList =  await navigator.mediaDevices.enumerateDevices();
    for (const mediaInfo of deviceList) {
      assert_not_equals(mediaInfo.kind, undefined, "mediaInfo's kind should exist.");
      assert_not_equals(mediaInfo.deviceId, "", "mediaInfo's deviceId should exist and not be empty.");
      assert_in_array(mediaInfo.kind, ["videoinput", "audioinput", "audiooutput"]);
    }
  }, "mediaDevices.enumerateDevices() is present and working - after capture");
</script>
</body>
</html>
